PV-50 Implement very basic authentication so can login as specific user
Implemented (via Django default behaviour) ability to login, logout, reset password etc.

Also modified pages which require a user to use current logged in user or a hard-coded default user.  This will need to be improved to work with sessions for non logged in users but this is fine for now.