PV-165 Implement common colors and style available to all users
Added makesuperuser command so that an admin user is created when creating a container.  Generates random password and prints it out - a good alternative to trying to store it somewhere secret.

Also added command to createsuperuser and add_common_data to entrypoint.sh which is executed whenever the container is created or updated.