[PV-241] Set up structure of unit tests (following re-think)
Ok all tests that can run are running.

A few have been removed as they tested code which is no longer part of the app (very sloppy) but we will need to re-write some of those tests to reflect the current codeset.

A couple have been skipped as they reflect functionality which isn't there yet (email secrets and proper mapping of input fields to output fields in all valid combinations)