[PV-261][WIP] File read chooses current active sheet which may not be the one with the data in it.
The logic in the view was incorrect, and didn't abort the transaction when there was an error in reading or parsing the plan, so we were left with orphan plan entries without plan activities and this created an error in manage plans view.

Introduced transaction.set_rollback(True) after a lot of trial and error and reading up on this.