[PV-109][WIP] Develop approach for deploying to staging
(37th attempt)

Getting there!

Include colour change of headings to check app deployment working (cyanish).

Have created a deploy-staging.yml file to drive the github action, which supersedes running a bash script.  I know it's not complete or correct but want to try pushing to Development branch to see how far it gets.

(NOTE: As I'm not running the deployment script the check for WIP isn't there at the moment.