[PV-109][WIP] Develop approach for deploying to staging
(38th attempt)

Realised that when running in Github actions it's superseding the docker-compose.yml file as the commands are directly in the actions yaml file.  So need to update yamle file so that it accesses the right Dockerfile and mounts the right volumes etc.  This is an interim attempt after some changes but not expecting this to work.