[PV-117][WIP] Combine configure activities and layout visual into one screen
This has been a massive set of changes so I'm not going to try to split them into logical commits.  This is still work in progress but have made loads of progress on re-structuring the API and the code which generates the rendered visual data so that it will make it easier to build the interactive main screen.  Also have done some work outstide the app to design the new screen - so will bring that in next - still as part of this ticket.